refactor: duplicate default.json5 into auto named preset
Renovate seems to think it should be using renovate.json as the default, but
what I actually want is renovate to also run in this repo, and default.json5 to
be the preset that gets used in this repo automatically.

To remove Renovate's confusion, it seems easier to be more specific about the
preset. This also allows for additional presets to be added.

I'm leaving the default.json5 for now for compatibility.